src/main.py
"""
THE BIG FAMILY LEGACY - AI Agents API
FastAPI application for AI processing workflows
"""

from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import Dict, List, Optional, Any
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Create FastAPI app
app = FastAPI(
    title="The Big Family Legacy - AI Agents",
    description="AI processing workflows for family legacy platform",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Request models
class MemoryAnalysisRequest(BaseModel):
    memory_id: str
    image_url: str
    user_description: Optional[str] = ""

class StoryGenerationRequest(BaseModel):
    person_id: str
    family_id: str
    tone: str = "emotional"
    length: str = "medium"

# Response models
class MemoryAnalysisResponse(BaseModel):
    success: bool
    memory_id: str
    ai_caption: str
    detected_faces: List[Dict]
    privacy_approved: bool

class StoryGenerationResponse(BaseModel):
    success: bool
    story_id: str
    story_content: str
    generation_time_ms: int

# Health check
@app.get("/health")
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "ai-agents",
        "version": "1.0.0"
    }

# Memory analysis endpoint
@app.post("/api/v1/analyze-memory", response_model=MemoryAnalysisResponse)
async def analyze_memory(request: MemoryAnalysisRequest):
    """
    Analyze uploaded memory (photo) with AI
    - Privacy check
    - Vision AI captioning
    - Face detection
    """
    try:
        # TODO: Implement actual workflow
        # For now, return placeholder
        return MemoryAnalysisResponse(
            success=True,
            memory_id=request.memory_id,
            ai_caption="A joyful family gathering with smiling faces",
            detected_faces=[],
            privacy_approved=True
        )
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Story generation endpoint
@app.post("/api/v1/generate-story", response_model=StoryGenerationResponse)
async def generate_story(request: StoryGenerationRequest):
    """
    Generate AI story for family member
    - Gather context from family tree
    - Curate relevant memories
    - Write biographical narrative
    - Privacy review
    """
    try:
        # TODO: Implement actual workflow
        # For now, return placeholder
        return StoryGenerationResponse(
            success=True,
            story_id="story-123",
            story_content="Once upon a time in a big family...",
            generation_time_ms=5000
        )
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# Debug endpoint
@app.get("/api/v1/debug")
async def debug_info():
    """Debug information"""
    return {
        "environment": {
            "supabase_url": os.getenv("SUPABASE_URL", "NOT_SET"),
            "hf_token_set": bool(os.getenv("HF_TOKEN")),
            "openai_key_set": bool(os.getenv("OPENAI_API_KEY"))
        },
        "agents_available": [
            "privacy_guard",
            "memory_curator",
            "genealogist",
            "biographer"
        ]
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
