requirements.txt
crewai==0.11.0
langgraph==0.2.0
langchain==0.1.0
langchain-openai==0.0.5
huggingface-hub==0.20.0
supabase==2.3.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.2.0
requests==2.31.0
